fix delete check car type before delete in car brand admin

+++ src/main/java/vn/edu/nlu/fit/datxedulich/services/BrandService.java
@@ -23,7 +23,11 @@ public class BrandService {
         brandDao.updateBrand(brand);
     }
 
-    public void deleteBrand(int id) {
+    public boolean deleteBrand(int id) {
+        if (brandDao.hasCarTypes(id)) {
+            return false;
+        }
         brandDao.deleteBrand(id);
+        return true;
     }
 }
